Fix eslint import extension
