+++ js/automation-editor/automation-editor.js
@@ -1,5 +1,5 @@
 import { h, render } from 'preact';
-import Automation from './automation';
+import Automation from './automation.js';
 
 window.AutomationEditor = function (mountEl, props, mergeEl) {
   return render(h(Automation, props), mountEl, mergeEl);
