+++ js/automation-editor/trigger/time.js
@@ -1,6 +1,6 @@
 import { h, Component } from 'preact';
 
-import { onChangeEvent } from '../../common/util/event';
+import { onChangeEvent } from '../../common/util/event.js';
 
 export default class TimeTrigger extends Component {
   constructor() {
