+++ js/automation-editor/trigger/trigger_edit.js
@@ -1,14 +1,14 @@
 import { h, Component } from 'preact';
 
-import EventTrigger from './event';
-import HassTrigger from './homeassistant';
-import MQTTTrigger from './mqtt';
-import NumericStateTrigger from './numeric_state';
-import StateTrigger from './state';
-import SunTrigger from './sun';
-import TemplateTrigger from './template';
-import TimeTrigger from './time';
-import ZoneTrigger from './zone';
+import EventTrigger from './event.js';
+import HassTrigger from './homeassistant.js';
+import MQTTTrigger from './mqtt.js';
+import NumericStateTrigger from './numeric_state.js';
+import StateTrigger from './state.js';
+import SunTrigger from './sun.js';
+import TemplateTrigger from './template.js';
+import TimeTrigger from './time.js';
+import ZoneTrigger from './zone.js';
 
 const TYPES = {
   event: EventTrigger,
