+++ js/automation-editor/trigger/zone.js
@@ -1,6 +1,6 @@
 import { h, Component } from 'preact';
 
-import { onChangeEvent } from '../../common/util/event';
+import { onChangeEvent } from '../../common/util/event.js';
 
 export default class ZoneTrigger extends Component {
   constructor() {
