+++ js/common/component/condition/numeric_state.js
@@ -1,6 +1,6 @@
 import { h, Component } from 'preact';
 
-import { onChangeEvent } from '../../util/event';
+import { onChangeEvent } from '../../util/event.js';
 
 export default class NumericStateCondition extends Component {
   constructor() {
