+++ js/common/component/condition/zone.js
@@ -1,6 +1,6 @@
 import { h, Component } from 'preact';
 
-import { onChangeEvent } from '../../util/event';
+import { onChangeEvent } from '../../util/event.js';
 
 export default class ZoneCondition extends Component {
   constructor() {
