+++ js/common/component/script/event.js
@@ -1,7 +1,7 @@
 import { h, Component } from 'preact';
 
-import JSONTextArea from '../json_textarea';
-import { onChangeEvent } from '../../util/event';
+import JSONTextArea from '../json_textarea.js';
+import { onChangeEvent } from '../../util/event.js';
 
 export default class EventAction extends Component {
   constructor() {
