+++ js/common/component/script/wait.js
@@ -1,5 +1,5 @@
 import { h, Component } from 'preact';
-import { onChangeEvent } from '../../util/event';
+import { onChangeEvent } from '../../util/event.js';
 
 export default class WaitAction extends Component {
   constructor() {
