+++ js/common/util/compute_state_domain.js
@@ -1,0 +1,9 @@
+import computeDomain from './compute_domain.js';
+
+export default function computeStateDomain(stateObj) {
+  if (!stateObj._domain) {
+    stateObj._domain = computeDomain(stateObj.entity_id);
+  }
+
+  return stateObj._domain;
+}
