+++ js/util.js
@@ -6,12 +6,12 @@
  * import these functions where we need them.
  */
 
-import attributeClassNames from './common/util/attribute_class_names';
-import computeDomain from './common/util/compute_domain';
-import computeStateDisplay from './common/util/compute_state_display';
-import formatDate from './common/util/format_date';
-import formatDateTime from './common/util/format_date_time';
-import formatTime from './common/util/format_time';
+import attributeClassNames from './common/util/attribute_class_names.js';
+import computeStateDomain from './common/util/compute_state_domain.js';
+import computeStateDisplay from './common/util/compute_state_display.js';
+import formatDate from './common/util/format_date.js';
+import formatDateTime from './common/util/format_date_time.js';
+import formatTime from './common/util/format_time.js';
 
 window.hassUtil = window.hassUtil || {};
 
@@ -21,7 +21,7 @@ const language = navigator.languages ?
 window.fecha.masks.haDateTime = window.fecha.masks.shortTime + ' ' + window.fecha.masks.mediumDate;
 
 window.hassUtil.attributeClassNames = attributeClassNames;
-window.hassUtil.computeDomain = computeDomain;
+window.hassUtil.computeDomain = computeStateDomain;
 window.hassUtil.computeStateDisplay = computeStateDisplay;
 window.hassUtil.formatDate = dateObj => formatDate(dateObj, language);
 window.hassUtil.formatDateTime = dateObj => formatDateTime(dateObj, language);
