+++ test-mocha/common/util/compute_domain.js
@@ -3,10 +3,10 @@ import computeDomain from '../../../js/common/util/compute_domain';
 const assert = require('assert');
 
 describe('computeDomain', () => {
-  it('Detects sensor domain', () => {
-    const stateObj = {
-      entity_id: 'sensor.test',
-    };
-    assert.strictEqual(computeDomain(stateObj), 'sensor');
+  it('Returns domains', () => {
+    assert.equal(computeDomain('sensor.bla'), 'sensor');
+    assert.equal(computeDomain('switch.bla'), 'switch');
+    assert.equal(computeDomain('light.bla'), 'light');
+    assert.equal(computeDomain('persistent_notification.bla'), 'persistent_notification');
   });
 });
