+++ test-mocha/common/util/compute_state_domain.js
@@ -1,0 +1,12 @@
+import computeStateDomain from '../../../js/common/util/compute_state_domain.js';
+
+const assert = require('assert');
+
+describe('computeStateDomain', () => {
+  it('Detects sensor domain', () => {
+    const stateObj = {
+      entity_id: 'sensor.test',
+    };
+    assert.strictEqual(computeStateDomain(stateObj), 'sensor');
+  });
+});
